Crate icu_provider_blob

icu_provider_blob contains BlobDataProvider, a BufferProvider implementation that supports loading data from a single serialized blob.

To build blob data, use the --format blob option of icu_provider_export:

$ icu4x-datagen --markers all --locales full --format blob --out data.postcard

Structs

BlobDataProvider

A data provider that reads from serialized blobs of data.