Crate icu_provider_blob

icu_provider_blob contains implementations of the ICU4X BufferProvider interface that load data from a single blob.

There are two exports:

BlobDataProvider supports data blobs loaded dynamically at runtime.

StaticDataProvider supports data blobs baked into the binary at compile time.

To build blob data, use the --format blob option of icu_datagen. For example, to build “hello world” data, run:

$ cargo run --features bin -p icu_datagen -- \
    --format blob \
    --hello-world-key \
    --all-locales \
    --out hello_world.postcard

For examples, see the specific data providers.

Structs

BlobDataProvider

A data provider loading data from blobs dynamically created at runtime.

StaticDataProvider

A data provider loading data statically baked in to the binary.