Crate icu_provider_blob

icu_provider_blob contains BlobDataProvider, a BufferProvider implementation that supports loading data from a single serialized blob.

To build blob data, use the --format blob2 option of icu_datagen:

$ icu4x-datagen --keys all --locales full --format blob2 --out data.postcard

You can also use --format blob if you need to support ICU4X versions prior to 1.4.

Structs

BlobDataProvider

A data provider that reads from serialized blobs of data.