Crate icu_provider_blob

icu_provider_blob contains implementations of the ICU4X DataProvider interface that load data from a single blob.

There are two exports:

BlobDataProvider supports data blobs loaded dynamically at runtime.

StaticDataProvider supports data blobs baked into the binary at compile time.

To build blob data, use the --format blob option of icu4x-datagen. For example, to build “hello world” data, run:

$ cargo run --bin=icu4x-datagen -- \
    --format blob \
    --hello-world-key \
    --all-locales \
    --out hello_world.postcard

Example

Create a StaticDataProvider from pre-built test data:

let _ = icu_testdata::get_static_provider();

For more examples, see the specific data providers.

Modules

export

Utilities for dumping data to an ICU4X filesystem tree.

Structs

BlobDataProvider

A data provider loading data from blobs dynamically created at runtime.

StaticDataProvider

A data provider loading data statically baked in to the binary.