Struct icu_provider_blob::BlobDataProvider

pub struct BlobDataProvider { /* private fields */ }

A data provider that reads from serialized blobs of data.

This enables data blobs to be read from arbitrary sources at runtime, allowing code and data to be separated. Alternatively, blobs can also be statically included at compile time.

BlobDataProvider implements BufferProvider, so it can be used in *_with_buffer_provider constructors across ICU4X.

Sync + Send

This provider uses reference counting internally. When the sync Cargo feature on the icu_provider crate is enabled, it uses Arc instead of Rc, making it Sync + Send.

Examples

Dynamic loading

Load “hello world” data from a postcard blob loaded at runtime:

use icu_locid::locale;
use icu_provider::hello_world::HelloWorldFormatter;
use icu_provider_blob::BlobDataProvider;
use writeable::assert_writeable_eq;

// Read an ICU4X data blob dynamically:
let blob = std::fs::read("tests/data/v2.postcard")
    .expect("Reading pre-computed postcard buffer");

// Create a DataProvider from it:
let provider = BlobDataProvider::try_new_from_blob(blob.into_boxed_slice())
    .expect("Deserialization should succeed");

// Check that it works:
let formatter = HelloWorldFormatter::try_new_with_buffer_provider(
    &provider,
    &locale!("la").into(),
)
.expect("locale exists");

assert_writeable_eq!(formatter.format(), "Ave, munde");

Static loading

Load “hello world” data from a postcard blob statically linked at compile time:

use icu_locid::locale;
use icu_provider::hello_world::HelloWorldFormatter;
use icu_provider_blob::BlobDataProvider;
use writeable::assert_writeable_eq;

// Read an ICU4X data blob statically:
const HELLO_WORLD_BLOB: &[u8] = include_bytes!("../tests/data/v2.postcard");

// Create a DataProvider from it:
let provider = BlobDataProvider::try_new_from_static_blob(HELLO_WORLD_BLOB)
    .expect("Deserialization should succeed");

// Check that it works:
let formatter = HelloWorldFormatter::try_new_with_buffer_provider(
    &provider,
    &locale!("la").into(),
)
.expect("locale exists");

assert_writeable_eq!(formatter.format(), "Ave, munde");

Implementations

impl BlobDataProvider

pub fn try_new_from_blob(blob: Box<[u8]>) -> Result<Self, DataError>

Create a BlobDataProvider from a blob of ICU4X data.

pub fn try_new_from_static_blob(blob: &'static [u8]) -> Result<Self, DataError>

Create a BlobDataProvider from a static blob. This is a special case of try_new_from_blob and is allocation-free.